mnl: update constructor to return io::Result not Nlmsg
This causes different behavior from the original in nlmsg_batch handling.
The original does not have buf size, but assumes its buf size is double of
the limit. Then it can return next buf event it exceeds the limit.

This wrapper does not assume a such, just holding its limit as same as the
buf size, so that it can not next if it exceeds the limit